frontend/requirements.txt
altair
pandas
streamlit
requests
python-dotenv
